updated code from int to long long

+++ Daily LeetCode Questions/1513. Number of substrings with only 1s.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int numSub(string s) {
-        int count1=0;
-        int result=0;
+        long long count1=0;
+        long long result=0;
         for(char c:s){
             if(c=='0'){
                 result+=(count1)*(count1+1)/2;
